+++ asc/src/commands/ls.rs
@@ -1,12 +1,13 @@
-use eyre::Result;
+use std::{env::current_dir, fs};
 
-use libasc::{filter_with_glob, repository::Repository};
-use relative_path::RelativePathBuf;
+use eyre::Result;
+use libasc::{change::FileChange, repository::Repository, utils::{filter_paths_with_glob_strict, hash_raw_bytes}};
+use relative_path::{PathExt, RelativePathBuf};
 
 #[derive(clap::Args)]
 pub struct Args {
     /// The pattern to glob against. Omitting this lists from the repository root.
-    patterns: Vec<String>,
+    patterns: Vec<RelativePathBuf>,
 
     /// Include hidden files.
     #[arg(short = 'a', long = "all")]
@@ -14,11 +15,18 @@ pub struct Args {
 
     /// List contents from another version. Omitting this uses the current version.
     #[arg(short, long)]
-    version: Option<String>
+    version: Option<String>,
+
+    /// List file change information.
+    #[arg(short = 'c')]
+    include_changes: bool
 }
 
-pub fn parse(args: Args) -> Result<()> {
+pub fn parse(mut args: Args) -> Result<()> {
     let repo = Repository::load()?;
+
+    let current_dir = current_dir()
+        .unwrap_or(repo.root_dir.clone());
 
     let snapshot = if let Some(raw_version) = args.version {
         let version = repo.normalise_version(&raw_version)?;
@@ -29,19 +37,25 @@ pub fn parse(args: Args) -> Result<()> {
         repo.fetch_current_snapshot()?
     };
 
-    let paths: Vec<RelativePathBuf> = snapshot.files
-        .into_keys()
-        .collect();
-
     if args.patterns.is_empty() {
-        for path in paths {
-            println!("{path}");
-        }
-
-        return Ok(());
+        args.patterns.push(RelativePathBuf::from("."));
     }
 
-    let mut valid_paths = filter_with_glob(args.patterns, &paths);
+    let filter_result = filter_paths_with_glob_strict(
+        &args.patterns,
+        &repo.staged_files,
+        &repo.root_dir
+    );
+
+    let mut valid_paths = match filter_result {
+        Ok(matches) => matches,
+        
+        Err(invalid_path) => {
+            eprintln!("Path outside of tree: {invalid_path}");
+
+            return Ok(());
+        }
+    };
 
     if valid_paths.is_empty() {
         eprintln!("No paths found.");
@@ -52,7 +66,32 @@ pub fn parse(args: Args) -> Result<()> {
     valid_paths.sort();
 
     for path in valid_paths {
-        println!("{path}");
+        let absolute = path.to_logical_path(&repo.root_dir);
+
+        let display_path = absolute.relative_to(&current_dir)?;
+
+        if !args.include_changes {
+            println!("{display_path}");
+
+            continue;
+        }
+        
+        if !absolute.exists() {
+            println!("{}", FileChange::Missing(display_path));
+
+            continue;
+        }
+
+        let data = fs::read(absolute)?;
+
+        let hash = hash_raw_bytes(data);
+
+        if hash == snapshot.files[path] {
+            println!("{}", FileChange::Unchanged(display_path));
+        }
+        else {
+            println!("{}", FileChange::Edited(display_path));
+        }
     }
 
     Ok(())